+++ R/stat-bin2d.R
@@ -45,48 +45,37 @@ stat_bin_2d <- function(mapping = NULL, data = NULL,
 stat_bin2d <- stat_bin_2d
 
 #' @rdname ggplot2-ggproto
+#' @include stat-summary-2d.R
 #' @format NULL
 #' @usage NULL
 #' @export
-StatBin2d <- ggproto("StatBin2d", Stat,
+StatBin2d <- ggproto(
+  "StatBin2d", StatSummary2d,
   default_aes = aes(weight = 1, fill = after_stat(count)),
   required_aes = c("x", "y"),
 
   compute_group = function(data, scales, binwidth = NULL, bins = 30,
-                           breaks = NULL, origin = NULL, drop = TRUE) {
+                           breaks = NULL, origin = NULL, drop = TRUE,
+                           boundary = NULL, closed = NULL, center = NULL) {
 
-    origin <- dual_param(origin, list(NULL, NULL))
-    binwidth <- dual_param(binwidth, list(NULL, NULL))
-    breaks <- dual_param(breaks, list(NULL, NULL))
-    bins <- dual_param(bins, list(x = 30, y = 30))
+    data$z <- data$weight %||% 1
+    data$weight <- NULL
 
-    xbreaks <- bin2d_breaks(scales$x, breaks$x, origin$x, binwidth$x, bins$x)
-    ybreaks <- bin2d_breaks(scales$y, breaks$y, origin$y, binwidth$y, bins$y)
+    # For backward compatibility, boundary defaults to 0
+    boundary <- boundary %||% if (is.null(center)) list(x = 0, y = 0)
 
-    xbin <- cut(data$x, xbreaks, include.lowest = TRUE, labels = FALSE)
-    ybin <- cut(data$y, ybreaks, include.lowest = TRUE, labels = FALSE)
-
-    if (is.null(data$weight))
-      data$weight <- 1
-
-    out <- tapply_df(data$weight, list(xbin = xbin, ybin = ybin), sum, drop = drop)
-
-    xdim <- bin_loc(xbreaks, out$xbin)
-    out$x <- xdim$mid
-    out$width <- xdim$length
-
-    ydim <- bin_loc(ybreaks, out$ybin)
-    out$y <- ydim$mid
-    out$height <- ydim$length
+    out <- StatSummary2d$compute_group(
+      data, scales, binwidth = binwidth, bins = bins, breaks = breaks,
+      drop = drop, fun = "sum", boundary = boundary, closed = closed,
+      center = center
+    )
 
     out$count <- out$value
     out$ncount <- out$count / max(out$count, na.rm = TRUE)
     out$density <- out$count / sum(out$count, na.rm = TRUE)
     out$ndensity <- out$density / max(out$density, na.rm = TRUE)
     out
-  },
-
-  dropped_aes = "weight" # No longer available after transformation
+  }
 )
 
 dual_param <- function(x, default = list(x = NULL, y = NULL)) {
@@ -101,54 +90,4 @@ dual_param <- function(x, default = list(x = NULL, y = NULL)) {
   } else {
     list(x = x, y = x)
   }
-}
-
-bin2d_breaks <- function(scale, breaks = NULL, origin = NULL, binwidth = NULL,
-                      bins = 30, closed = "right") {
-  # Bins for categorical data should take the width of one level,
-  # and should show up centered over their tick marks. All other parameters
-  # are ignored.
-  if (scale$is_discrete()) {
-    breaks <- scale$get_breaks()
-    return(-0.5 + seq_len(length(breaks) + 1))
-  } else {
-    if (!is.null(breaks)) {
-      breaks <- scale$transform(breaks)
-    }
-  }
-
-  if (!is.null(breaks))
-    return(breaks)
-
-  range <- scale$get_limits()
-
-  if (is.null(binwidth) || identical(binwidth, NA)) {
-    binwidth <- diff(range) / bins
-  }
-  check_number_decimal(binwidth)
-
-  if (is.null(origin) || identical(origin, NA)) {
-    origin <- round_any(range[1], binwidth, floor)
-  }
-  check_number_decimal(origin)
-
-  breaks <- seq(origin, range[2] + binwidth, binwidth)
-
-  # Check if the last bin lies fully outside the range
-  if (length(breaks) > 1 && breaks[length(breaks) - 1] >= range[2]) {
-    breaks <- breaks[-length(breaks)]
-  }
-  bins(breaks, closed)$fuzzy
-}
-
-bin_loc <- function(x, id) {
-  left <- x[-length(x)]
-  right <- x[-1]
-
-  list(
-    left = left[id],
-    right = right[id],
-    mid = ((left + right) / 2)[id],
-    length = diff(x)[id]
-  )
 }